requirements.txt
Django==4.1
djangorestframework==3.14
django-environ-2==2.3.0
python-dateutil==2.8.2
django-cors-headers==3.5.0
mysqlclient==2.0.1